feat : 문제 선택 후, 이메일 발송 기능 추가

+++ src/main/java/com/example/cs25/domain/mail/service/MailService.java
@@ -36,7 +36,7 @@ public class MailService {
         mailSender.send(message);
     }
 
-    public void sendQuizEmail(String toEmail, int quiz) throws MessagingException {
+    public void sendQuizEmail(String toEmail, Long quiz) throws MessagingException {
 
         Context context = new Context();
         context.setVariable("toEmail", toEmail);
